ADD: Initial version of trees.bmp
Based on Crackd's original sources, so still "data-driven".
The source was an estimate (calculated by a super-computer) for forest cover in 1066, actually broken down to several types of trees (coniferous, broadleave, ...).
Will need adjustment because the base was quite "blocky" and some trees are now misplaced or in the water...