TWEAK: Did a performance pass
Worked out okayish.
Graphics demand still surges in areas with pine trees especially (these must be really unoptimized - I mean, since they all look the same they should not take up as many draw calls as they seem to).
So Siberia needs to be deforested some more.
India on the other hand is now a little too tree-less, so in areas of jungle it should maybe be increased a bit.