Assets: Shoreline fixed North Africa & Marsh terrain update
Shoreline fixed from Israel to Strait of Gibraltar, including all major Chotts and Sabkhas (salt lakes).
The marsh terrain WIP was updated with data from the PEATMAP project.
(These are new files instead of updates to the old ones, because I have grown used to doing this locally and won't bother to change it for the internetz.)